Moves util/log.go to util/log/log.go
