+++ util/log/BUILD.bazel
@@ -1,0 +1,9 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_library")
+
+go_library(
+    name = "log",
+    srcs = ["log.go"],
+    importpath = "github.com/OffchainLabs/bold/util/log",
+    visibility = ["//visibility:public"],
+    deps = ["@com_github_ethereum_go_ethereum//log"],
+)
