+++ util/log/log_test.go
@@ -1,4 +1,4 @@
-package util
+package log
 
 import (
 	"errors"
